requirements.txt
ultralytics
opencv-python
pyserial
fastapi
uvicorn
pydantic
requests
streamlit
pandas
plotly
google-generativeai
numpy
traci
